+++ tests/testsuite/support/mod.rs
@@ -137,6 +137,8 @@ pub mod git;
 pub mod paths;
 pub mod publish;
 pub mod registry;
+#[macro_use]
+pub mod resolver;
 
 /*
  *
